requirements.txt
Flask==2.3.2
Flask-SocketIO==5.3.6
requests==2.32.3
eventlet==0.33.3
dnspython==2.3.0
torch==2.9.0
torchcodec==0.8
pyannote.audio==4.0.1
soundfile==0.13.1
ffmpeg-python==0.2.0
gunicorn==23.0.0
pydub==0.25.1
